requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
pydantic==2.12.5
sqlalchemy==2.0.23
python-dotenv==0.19.2
asyncpg==0.29.0
email-validator==2.1.0
mangum==0.17.0
PyJWT==2.10.1
bcrypt==4.0.1